Move definitions of BenchResult and downstream dependencies to separate file
Similar for JsonCache, and ProcessRunner.
